Added a CMakeLists.txt file for compiling bind
This allows for options such as '-DUSE_F2C' to turn on
or off the useage of F2C. If F2C is on (default), it must be able
to link to F2C libraries. If F2C is off, it must be able
to find a suitable fortran compiler. It also allows one
to define the F2C library location with '-DF2C_LIB='. I've
tested it with F2C and without, and both seem to work.
In addition, one can use '-DPARM_FILE_LOC=' to set the
parameter file location. Otherwise, it just uses the default
one in the tightbind directory. All of this info is printed
when cmake is run. Recommended usage: when in tightbind directory,
type 'mkdir build; cd build; cmake ..; make'. There are a lot
of warnings, but it creates the executable.