Added default eht params in case file not found (#6)
Default eht parameters are in eht_parms.h. If the param file
cannot be opened, it prints that the file was unable to be
opened and it reads data from eht_parms.h instead. This commit
also fixes a segmentation fault when the param file was not
found because functions like rewind() and skipcomments() are
no longer called on the param file if it doesn't exist.