Added options for linking to blas and lapack
When running cmake, the user can use the definition
-DUSE_BLAS_LAPACK=ON to link to blas and lapack for matrix
multiplication (instead of using cboris and diag subroutines
in yaehmop). It uses cmake to find blas and lapack and links
to them dynamically by default. If the user specifies
-DSTATIC_BLAS_LAPACK=ON as well, it will link to them
statically. I've tried every combination of those two definitions
with -DUSE_F2C=<ON/OFF>, and I've ran tests on diamond and they
all appear to work.

In addition, I added BUILD_STATIC as an option.
BUILD_STATIC=ON means that a static executable will be built.
There are some issues with BUILD_STATIC=ON and USE_F2C=ON, and I
think it's because f2c has some dynamic dependencies. So the
CMakeLists.txt does not allow both to be used together. Every
other combination (using blas and lapack, for instance) seems
to work just fine.