Added an option to use stdin and stdout (#9)
If, in the place of the input file argument the user says
"--use_stdin_stdout", then the program will use stdin to read
the input, and it will print all output to stdout. This is great for
programs running bind as a subprocess. Also, one line is printed to
the end of the band data to indicate a completion of the band data.