Resize char array to prevent crashing on Mac (#12)
For some reason, when reading the default parameters that I
added, Mac will crash since there is not enough memory for the
strcpy, but Linux does not. I think we can allow for a little
more flexibility for the instring size. A size 240 should not
have a noticeable impact on performance. This fixes the Mac
issue I was noticing.