Changed an ordering to fix a reallocation bug (#15)
The ordering was changed so a reallocation is only performed
if truly necessary (before, a reallocation would be performed
as soon as the max size was hit - regardless of whether it was
actually needed). Now, a reallocation is only performed if we are
to add another and there is not enough room.

This fixes a bug on Windows that causes crashing. The
reallocation function appears to be the culprit. It is
currently unknown what is wrong with it. Thus, if one sees
crashing again on Windows, this is a place where they should
look for a potential cause.