Updated C, Na, and Si parameters with fit data
We used Daniel Frederickson's EHT fitting program here:

http://www.chem.wisc.edu/~danny/software/ehtuner/

Stacey, T. E. and Frederickson, D. C., Dalton Trans., 2012,41, 7801-7813

to fit C, Na, and Si parameters to VASP data. To be more specific,
we used PBE for C and Na, and HSE06 for Si (we used HSE06 for Si because
it has a band gap and HSE06 does a better job at calculating band gaps).
In addition, we only fit the coefficients and exponents (we did NOT
fit the ionization potentials). We also added some d parameters to Si.

We used graphite for C, and just the ambient solid crystals for Na and
Si.

These parameters seem to give us much better band structures. For
example, the old parameters of Na gave a strange rise at the gamma
point which is not reasonable. The new parameters have eliminated
that issue.

For Si, the old parameters gave a direct band gap. The new parameters
give a correct indirect band gap.

For C-graphite, there weren't any major features that needed to change,
but we thought we would fit it too since we were fitting the others anyways.

Overall, I think these parameters result in a nice improvement.